Further improve package script wrapper
