+++ scripts/package.py
@@ -15,14 +15,8 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-import sys
-import os
-
-sys.path.insert(
-    0, os.path.abspath(os.path.join(os.path.dirname(__file__), "../deskflow/scripts"))
-)
-
-import package
+import os, argparse
+import lib.bootstrap as bootstrap
 
 PRODUCT_NAME = "Synergy"
 DESKFLOW_SOURCE_DIR = "deskflow"
@@ -30,18 +24,37 @@ DESKFLOW_BUILD_DIR = "build/deskflow"
 DEFAULT_PREFIX = "synergy"
 DIST_DIR = "dist"
 VERSION_FILENAME = "deskflow/VERSION"
+TEST_CMD = "synergy-server --version"
+PACKAGE_NAME = "synergy"
 
 
 def main():
+    parser = argparse.ArgumentParser()
+    parser.add_argument(
+        "--leave-test-installed",
+        action="store_true",
+        help="Leave test package installed",
+    )
+    args = parser.parse_args()
+
+    bootstrap.ensure_in_venv(__file__)
+    bootstrap.add_deskflow_path(__file__)
+
+    import package
+
     filename_base = get_env("SYNERGY_PACKAGE_PREFIX", default=DEFAULT_PREFIX)
     version = package.get_app_version(VERSION_FILENAME)
+
     package.package(
         filename_base,
         version,
         DESKFLOW_BUILD_DIR,
         DIST_DIR,
+        TEST_CMD,
         PRODUCT_NAME,
+        PACKAGE_NAME,
         source_dir=DESKFLOW_SOURCE_DIR,
+        leave_test_installed=args.leave_test_installed,
     )
